move condition tree into subcomponent and style panel headers
