+++ src/interface/src/app/map/map-control-panel/map-control-panel.component.scss
@@ -1,3 +1,11 @@
+@use 'sass:map';
+@use '@angular/material' as mat;
+
+@import "../../../styles.scss";
+
+$color-config:    mat.get-color-config($planscape-frontend-theme);
+$primary-palette: map.get($color-config, 'primary');
+
 .controls-container {
   background-color: white;
   box-sizing: border-box;
@@ -58,6 +66,7 @@
 .layer-control-container {
   display: flex;
   flex-direction: row;
+  margin: 5px;
 }
 
 .layer-controls-tab-group {
@@ -72,43 +81,28 @@
   margin: 15px 0;
 }
 
-.layer-radio-button {
-  margin: 5px;
-}
-
 .layer-checkbox {
-  margin: 0px 5px 0px;
+  margin: 15px 0;
 }
 
-.condition-tree-invisible {
-  display: none;
+.layer-panel-header {
+  background-color: #D8D8D8 !important;
+  font-size: 14px;
+  font-weight: 700;
+  height: 40px !important;
+  line-height: 20px;
 }
 
-.condition-tree ul,
-.condition-tree li {
-  margin-top: 0;
-  margin-bottom: 0;
-  list-style-type: none;
+// Overrides of Material component styles
+
+::ng-deep .mat-radio-checked {
+  label {
+    color: mat.get-color-from-palette($primary-palette, 500);
+  }
 }
 
-/*
- * This padding sets alignment of the nested nodes.
- */
-.condition-tree .mat-nested-tree-node div[role=group] {
-  padding-left: 20px;
-}
-
-/*
- * Padding for leaf nodes.
- * Leaf nodes need to have padding so as to align with other non-leaf nodes
- * under the same parent.
- */
-.condition-tree div[role=group] > .mat-tree-node {
-  padding-left: 20px;
-}
-
-.mat-tree-node > .disabled {
-  opacity: 0.5;
+::ng-deep .mat-expansion-panel-body {
+  padding-bottom: 0px !important;
 }
 
 // Workaround for unresolved Angular bug re: tab width on desktop
